temp

+++ apps/dashboards/tests/test_views.py
@@ -63,6 +63,8 @@ class HelpfulVotesGraphTests(TestCase):
         self.group = group(name='Contributors', save=True)
         try:
             self.redis = redis_client('helpfulvotes')
+            if self.redis is None:
+                raise SkipTest
             self.redis.flushdb()
             self.REDIS_KEY = settings.HELPFULVOTES_UNHELPFUL_KEY
         except (KeyError, AttributeError):
